Enum kay::Fate

pub enum Fate {
    Live,
    Die,
}

Return type of message handling functions, signifying if an Actor/SubActor should live on after receiving a certain message type.

Note: so far only has an effect on SubActors in Swarms

Variants
Live

Means: the Actor/SubActor should live on

Die

Means: the Actor/SubActor should be stopped, its state can be deallocated