Struct kay::ActorDefiner

pub struct ActorDefiner<'a, A> { /* fields omitted */ }

Helper that is used to define actor behaviour (message handlers).

It is passed to the closure arguments of ActorSystem::add and ActorSystem::extend

Methods

impl<'a, A: 'static> ActorDefiner<'a, A>

fn on<M: Message, F>(&mut self, handler: F) where
F: Fn(&M, &mut A, &mut World) -> Fate + 'static,

Attach a new message handler to an actor, defined by a closure which will receive 3 arguments:

the received message (can conveniently be destructured already as an argument)

the current actor state that can be mutated

a World to identify and send messages to other actors

the_counter.on(|&IncrementCount { increment }, counter, world| {
    counter.count += increment;
    world.send_to_id_of::<Logger, _>(format!("New count: {}", counter.count));
});

fn on_critical<M: Message, F>(&mut self, handler: F) where
F: Fn(&M, &mut A, &mut World) -> Fate + 'static,

Same as on but continues to receive after the ActorSystem panicked.

fn world(&mut self) -> World

Access a World of the system that an actor is being defined in, can be used to identify actors (and keep the ID) or send messages at define-time.

let logger_id = the_counter.world().id::<Logger>();

the_counter.on(move |&IncrementCount { increment }, counter, world| {
    counter.count += increment;
    world.send(logger_id, format!("New count: {}", counter.count));
});

the_counter.world().send(logger_id, "Just defined Counter!");

fn on_packet<M: Message, F>(&mut self, handler: F, critical: bool) where
F: Fn(&Packet<M>, &mut A, &mut World) -> Fate + 'static,

Advanced: Can be used to register a handler not only for a message, but for a whole packet (precise recipient id + message), in case a particular sub-actor needs to be identified in the handler, like Swarm does.