Struct kay::ActorSystem

pub struct ActorSystem { /* fields omitted */ }

The main thing inside of which all the magic happens.

An ActorSystem contains the states of all registered actors, message inboxes (queues) for each registered actor, and message dispatchers for each registered (Actor,Message) pair.

It can be controlled from the outside to do message passing and handling in turns.

Methods

impl ActorSystem

fn new(panic_callback: Box<Fn(Box<Any>, &mut World)>) -> ActorSystem

Create a new ActorSystem (usually only one per application is needed). Expects to get a panic callback as a parameter that is called when an actor panics during message handling and can thus be used to for example display the panic error message.

Note that after an actor panicking, the whole ActorSystem switches to a panicked state and only passes messages anymore which have been marked as critically received using ActorDefiner::on_critical.

fn add<A: 'static, D: Fn(ActorDefiner<A>)>(&mut self, actor: A, define: D)

Add a new actor to the system given an initial actor state and a closure that takes an ActorDefiner to define message handlers for this actor.

system.add(Logger::new(), |mut the_logger| {
    the_logger.on(|&Message { text }, logger, world| {
        //...
    });
});

fn extend<A: 'static, D: Fn(ActorDefiner<A>)>(&mut self, define: D)

Extend a previously added actor using a closure that takes an ActorDefiner to define additional message handlers for this actor.

This is useful if you want to implement different aspects of an actor in different files.

system.extend::<Logger, _>(|mut the_logger| {
    the_logger.on(|_: &ClearAll, logger, world| {
        //...
    });
});

fn send<M: Message>(&mut self, recipient: ID, message: M)

Send a message to the actor with a given ID. This is only used to send messages into the system from outside. Inside actor message handlers you always have access to a World that allows you to send messages.

fn id<A2: 'static>(&mut self) -> ID

Get the ID of a previously added actor. This is only used to identify actors from the outside. Inside actor message handlers you always have access to a World that allows you to identify actors.

fn process_all_messages(&mut self)

Processes all sent messages, and messages which are in turn sent during the handling of messages, up to a recursion depth of 1000.

This is typically called in the main loop of an application.

By sending different "top-level commands" in to the system and calling process_all_messages inbetween, different aspects of an application (for example, UI, simulation, rendering) can be run isolated from each other, in a fixed order of "turns" during each main-loop iteration.