Struct kay::World

pub struct World(_);

Gives limited access to an ActorSystem (typically from inside, in a message handler) to identify other actors and send messages to them.

Methods

impl World

fn send<M: Message>(&mut self, receiver: ID, message: M)

Send a message to a (sub-)actor with the given ID.

world.send(child_id, Update {dt: 1.0});

fn id<A2: 'static>(&mut self) -> ID

Identify an actor based on type.

let logger_id = world.id::<Logger>();

fn send_to_id_of<A: 'static, M: Message>(&mut self, message: M)

Shorthand for identifying an actor and then sending a message to it.

world.send_to_id_of::<Logger, _>("New message!");
// is equivalent to
let logger_id = world.id::<Logger>();
world.send(logger_id, "New message!");

fn broadcast_to_id_of<A: 'static, M: Message>(&mut self, message: M)

Shorthand to broadcast something to all subactors of a Swarm.

world.broadcast_to_id_of::<UIElement, _>(UpdateUI);
// is equivalent to
let all_elements = world.id::<Swarm<UIElement>>().broadcast();
world.send(all_elements, UpdateUI);