Module kay::swarm

Tools for dealing with large amounts of identical actors

Structs

Create
A message for adding a new sub-actor to a Swarm given its initial state.

CreateWith
A message for adding a new sub-actor to a Swarm given its initial state and an initial message that the sub-actor will handle immediately after creation.

SubActorDefiner
Helper that is used to define behaviour (message handlers) of sub-actors in a Swarm. Analogous to ActorDefiner but with some Swarm-related extras.

Swarm
A container-like actor, housing many sub-actors of identical behaviour.

ToRandom
A wrapper for messages to send a message to a random subset of the sub-actors of a Swarm.

Traits

StorageAware
Trait that allows dynamically sized SubActors to provide a "typical size" hint to optimize their storage in a Swarm

SubActor
Trait that sub-actors of a Swarm have to implement so their internally stored sub-actor ID can be gotten and set.