Struct kay::swarm::SubActorDefiner

pub struct SubActorDefiner<'a, SA: 'static>(_);

Helper that is used to define behaviour (message handlers) of sub-actors in a Swarm. Analogous to ActorDefiner but with some Swarm-related extras.

It is passed to the closure argument of Swarm::subactors.

Methods

impl<'a, SA: SubActor + 'static> SubActorDefiner<'a, SA>

fn on<M: Message, F>(&mut self, handler: F) where
F: Fn(&M, &mut SA, &mut World) -> Fate + 'static,

Analogous to ActorDefiner::on, the closure argument is passed:

the received message (can conveniently be destructured already as an argument)

the current sub-actor state that can be mutated

a World to identify and send messages to other actors

fn on_critical<M: Message, F>(&mut self, handler: F) where
F: Fn(&M, &mut SA, &mut World) -> Fate + 'static,

Same as on but continues to receive after the ActorSystem panicked. (Analogous to ActorDefiner::on_critical)

fn on_create_with<M: Message, F>(&mut self, handler: F) where
F: Fn(&M, &mut SA, &mut World) -> Fate + 'static,

Allows sub-actors to be created with a particular message (the one handled by the given closure) as an initial message.

This can then be triggered by sending a CreateWith message to the swarm.

fn on_random<M: Message, F>(&mut self, handler: F) where
F: Fn(&M, &mut SA, &mut World) -> Fate + 'static,

Allows a message (the one handled by the given closure) to be received by a random subset of the sub-actors of a Swarm.

This can then be triggered by sending a ToRandom message to the swarm.

fn world(&mut self) -> World

Access a World of the system that a sub-actor is being defined in, can be used to identify actors (and keep the ID) or send messages at define-time.

(Analogous to ActorDefiner::world)

fn on_packet<M: Message, F>(&mut self, handler: F, critical: bool) where
F: Fn(&Packet<M>, &mut SA, &mut World) -> Fate + 'static,

Advanced: Can be used to register a handler not only for a message, but for a whole packet (precise recipient id + message).

(Analogous to ActorDefiner::on_packet)