Struct kay::swarm::Swarm

pub struct Swarm<SubActor> { /* fields omitted */ }

A container-like actor, housing many sub-actors of identical behaviour.

Offers efficient storage of and broadcasting to its sub-actors.

New sub-actors can be added to a swarm using Create or CreateWith.

Methods

impl<SA: SubActor> Swarm<SA>

fn new() -> Self

Create an empty Swarm.

fn subactors<S>(subactor_definition: S) -> impl Fn(ActorDefiner<Self>) where
S: Fn(SubActorDefiner<SA>) + 'static,

Define message handlers for subactors of a Swarm when it is added to a system or extended. See SubActorDefiner to see what can be defined.

system.add(Swarm::<UIElement>::new(), Swarm::<UI>::subactors(|mut each_element| {
   each_element.on(|&Update, element, world| {
       //...
   });
}));