Trait kay::swarm::RecipientAsSwarm

pub trait RecipientAsSwarm<M: Message>: Sized {
    fn receive(_swarm: &mut Swarm<Self>, _message: &M) -> Fate { ... }
    fn receive_packet(swarm: &mut Swarm<Self>, packet: &Packet<M>) -> Fate { ... }
}

Helper trait for dispatching messages either to SubActors of a Swarm or the Swarm itself.

By implementing this trait on a SubActor, you indirectly control what Swarm<SubActor> will be a Recipient of. This is required since you can't implement Recipient on Swarm yourself, both being foreign types.

The default implementation will make Swarm<SubActor> a Recipient<Message> for all messages that SubActor is a Recipient of, dispatching incoming messages to the correct sub-actor based on the sub-actor id part of the Packet's recipient field.

To override this behaviour and handle all messages of a certain type on the Swarm level, you can implement RecipientAsSwarm<MessageForSwarm> for SubActor yourself.

Further, already given implementations define how a Swarm handles RequestConfirmation, ToRandom, Create and CreateWith messages.

Provided Methods

fn receive(_swarm: &mut Swarm<Self>, _message: &M) -> Fate

Analogous to Recipient::receive

fn receive_packet(swarm: &mut Swarm<Self>, packet: &Packet<M>) -> Fate

Analogous to Recipient::receive_packet