Trait kay::swarm::SubActor

pub trait SubActor: Compact + StorageAware + 'static {
    fn id(&self) -> ID;
    unsafe fn set_id(&mut self, id: ID);
}

Trait that sub-actors of a Swarm have to implement so their internally stored sub-actor ID can be gotten and set.

Furthermore, a SubActor has to implement Compact, so a Swarm can compactly store each SubActor's potentially dynamically-sized state.

This trait can be auto-derived on structs that contain a field _id: Option<ID> using kay_macros

Required Methods

fn id(&self) -> ID

Get the full ID (Swarm type id + sub-actor id) of self

unsafe fn set_id(&mut self, id: ID)

Set the full ID (Swarm type id + sub-actor id) of self (called internally by Swarm)