Trait kay::TypedID

pub trait TypedID: Copy + Clone + Sized + Debug + Hash {
    fn as_raw(&self) -> RawID;

    unsafe fn from_raw(raw: RawID) -> Self;
}

TypedID is a construct on top of a RawID that can refer to a specific kind of actor, or actor trait at compile time

Required Methods

fn as_raw(&self) -> RawID

Get the underlying RawID

unsafe fn from_raw(raw: RawID) -> Self

Construct a new TypedID from a RawID - this implies knowledge about the type of actor referenced by the RawID