Struct kay::ActorSystem

pub struct ActorSystem { /* fields omitted */ }

An ActorSystem contains the states of all registered actors, message inboxes (queues) for each registered actor, and message handlers for each registered (Actor,Message) pair.

Methods

impl ActorSystem

fn create_the_system(panic_callback: Box<Fn(Box<Any>)>) -> Box<ActorSystem>

Create the global singleton instance of ActorSystem. Sets THE_SYSTEM. Expects to get a panic callback as a parameter that is called when an actor panics during message handling and can thus be used to for example display the panic error message.

Note that after an actor panicking, the whole ActorSystem switches to a panicked state and only passes messages anymore which have been marked as critical using Actor::handle_critically.

fn send<M: Message>(&mut self, recipient: ID, message: M)

Send a message to the actor with a given ID. This is usually never called directly, instead, ID << Message is used to send messages.

fn process_all_messages(&mut self)

Processes all sent messages, and messages which are in turn sent during the handling of messages, up to a recursion depth of 1000.

This is typically called in the main loop of an application.

By sending different "top-level commands" in to the system and calling process_all_messages inbetween, different aspects of an application (for example, UI, simulation, rendering) can be run isolated from each other, in a fixed order during each loop iteration.

fn short_id<A: Actor>(&self) -> ShortTypeId

Get the short type id for an Actor. Ususally never called directly, use Actor::id() instead to get the full ID of a registered Actor.