Struct kay::World

pub struct World(_);

Gives limited access to an ActorSystem (typically from inside, in a message handler) to identify other actors and send messages to them.

Methods

impl World

pub fn send<M: Message>(&mut self, receiver: ID, message: M)

Send a message to a (sub-)actor with the given ID.

world.send(child_id, Update {dt: 1.0});

pub fn local_first<A: Actor>(&mut self) -> ID

Get the ID of the first machine-local instance of an actor.

pub fn global_first<A: Actor>(&mut self) -> ID

Get the ID of the first instance of an actor on machine 0

pub fn local_broadcast<A: Actor>(&mut self) -> ID

Get the ID for a broadcast to all machine-local instances of an actor.

pub fn global_broadcast<A: Actor>(&mut self) -> ID

Get the ID for a global broadcast to all instances of an actor on all machines.

pub fn allocate_instance_id<A: 'static + Actor>(&mut self) -> ID

Synchronously allocate a instance id for a instance that will later manually be added to a Swarm

pub fn local_machine_id(&mut self) -> u8

Get the id of the machine that we're currently in