Module kay::swarm

Tools for dealing with large amounts of identical actors

Structs

Confirmation
The reply to a RequestConfirmation

Create
A message for adding a new sub-actor to a Swarm given its initial state.

CreateWith
A message for adding a new sub-actor to a Swarm given its initial state and an initial message that the sub-actor will handle immediately after creation.

RequestConfirmation
A wrapper for messages to request a confirmation about message receival.

Swarm
Offers efficient storage and updating of large numbers of identical SubActors and is typically used whenever there is more than one actor with the same type/behaviour.

ToRandom
A wrapper for messages to send a message to random sub-actors.

Traits

RecipientAsSwarm
Helper trait for dispatching messages either to SubActors of a Swarm or the Swarm itself.

StorageAware
Trait that allows dynamically sized SubActors to provide a "typical size" hint to optimize their storage in a Swarm

SubActor
Trait that sub-actors of a Swarm have to implement in order for the Swarm to set the sub-actor id and the sub-actor to get its own id in a standardized way.