Struct kay::swarm::Swarm

pub struct Swarm<SubActor> { /* fields omitted */ }

Offers efficient storage and updating of large numbers of identical SubActors and is typically used whenever there is more than one actor with the same type/behaviour.

If SubActor can receive Message, then Swarm<SubActor> can do so as well, redirecting each such message to the correct sub-actor by default (see RecipientAsSwarm)

Methods

impl<SA: SubActor> Swarm<SA>

fn new() -> Self

Create an empty Swarm

fn all() -> ID where Self: Sized

Get an ID that refers to all sub-actors inside a Swarm. If you send a message to this ID, all sub-actors will receive it (Exception: the Swarm handles the message itself, see RecipientAsSwarm)