Trait kay::Actor

pub trait Actor: 'static + Sized {
    fn register_with_state(initial_state: Self) { ... }
    fn register_default() where Self: Default { ... }
    fn id() -> ID { ... }
    fn handle<M: Message>() where Self: Recipient<M> { ... }
    fn handle_critically<M: Message>() where Self: Recipient<M> { ... }
}

Any Rust data structure can become an actor by:

Deriving Actor

Deriving Recipient<M> for all message types that this actor will handle

Registering itself and all handled messages with the actor system

Provided Methods

fn register_with_state(initial_state: Self)

Register an Actor with the system, using the given instance of the actor type as the initial state for the actor.

After this call, the Actor has an ID (see id()) and can receive messages (if according message handlers have been registered with handle or handle_critically)

fn register_default() where Self: Default

Like register_with_state, using the default value for Actors that implement Default

fn id() -> ID

Get the ID of an Actor (only works after the actor has been registered)

fn handle<M: Message>() where Self: Recipient<M>

Register a message handler for a message M for an Actor, which is defined in the implementation of `Recipient for this actor.

fn handle_critically<M: Message>() where Self: Recipient<M>

Like handle, but marks this message type as critical for Actor, which means that this actor will still receive such messages even after a panic occured in the ActorSystem

This makes it possible to keep for example rendering, debug message display and camera movement alive, even after an otherwise fatal panic.